Fix wrong import of shots_service
